moved gtag call from main.jsx to root
